Fix documentation scripts
