+++ package.json
@@ -39,7 +39,7 @@
     "clean:dist": "rimraf temp && rimraf dist",
     "clean:dist:x86_64": "rm -rf temp && rm -rf dist",
     "clean:docs": "rimraf app/docs && rimraf docs/ids-website",
-    "documentation": "npm run clean:docs && node ./scripts/deploy-documentation.js",
+    "documentation": "node ./scripts/deploy-documentation.js",
     "eslint": "npx eslint src tests app",
     "eslint:error-only": "npx eslint --quiet src tests app",
     "eslint:log": "npx eslint -o eslint.log src tests app",
@@ -82,19 +82,19 @@
     "jquery": "^3.7.1"
   },
   "devDependencies": {
-    "@axe-core/playwright": "^4.9.1",
+    "@axe-core/playwright": "^4.10.0",
     "@babel/core": "^7.25.2",
     "@babel/eslint-parser": "^7.25.1",
     "@babel/plugin-external-helpers": "^7.24.7",
-    "@babel/preset-env": "^7.25.3",
-    "@percy/cli": "^1.29.1",
+    "@babel/preset-env": "^7.25.4",
+    "@percy/cli": "^1.29.2",
     "@percy/playwright": "^1.0.6",
-    "@percy/sdk-utils": "^1.29.1",
-    "@playwright/test": "^1.46.0",
+    "@percy/sdk-utils": "^1.29.2",
+    "@playwright/test": "^1.46.1",
     "@rollup/plugin-babel": "^6.0.4",
     "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-node-resolve": "^15.2.3",
-    "@types/node": "^22.1.0",
+    "@types/node": "^22.5.0",
     "archiver": "^7.0.1",
     "babel-loader": "^9.1.3",
     "babel-plugin-transform-optional-chaining": "^7.0.0-beta.3",
@@ -107,7 +107,7 @@
     "eslint-plugin-babel": "^5.3.1",
     "eslint-plugin-compat": "6.0.0",
     "eslint-plugin-import": "^2.29.1",
-    "eslint-plugin-jasmine": "^4.2.0",
+    "eslint-plugin-jasmine": "^4.2.1",
     "eslint-plugin-jasmine-jquery": "^1.0.0",
     "express": "^4.19.2",
     "express-csp": "^0.1.3",
@@ -122,23 +122,23 @@
     "ids-css": "^1.5.0",
     "js-yaml": "^4.1.0",
     "markdownlint-cli": "^0.41.0",
-    "marked": "^13.0.3",
+    "marked": "^14.0.0",
     "mmm": "^0.2.2",
     "nodemon": "^3.1.4",
     "nyc": "^17.0.0",
     "r2": "^2.0.1",
     "release-it": "^17.6.0",
     "rimraf": "^6.0.1",
-    "rollup": "^4.20.0",
+    "rollup": "^4.21.0",
     "rollup-plugin-analyzer": "^4.0.0",
     "rollup-plugin-istanbul": "^5.0.0",
     "rollup-plugin-license": "^3.5.2",
     "sass": "1.77.8",
     "slash": "^5.1.0",
-    "stylelint": "^16.8.1",
+    "stylelint": "^16.8.2",
     "stylelint-config-standard-scss": "^13.1.0",
     "stylelint-scss": "^6.5.0",
-    "terser": "^5.31.3",
+    "terser": "^5.31.6",
     "vinyl": "^3.0.0",
     "vinyl-fs": "^4.0.0",
     "yargs": "^17.7.2",
